updated authorship
(This is a Partial RECONSTRUCTION of this change, based on old log
data.  It preserves the name of the user who made the change, the date
of the change, the old log message, and some (but not all) of the change data.
The RCS file backing up this CVS file was destroyed during 
hardware trouble around May, 2003. 

   --Steve Augart, Wednesday, August 13, 2003)