rvm/src/vm/libSupport/java/lang/reflect VMArray.java,1.2,1.3 Update of /cvsroot/jikesrvm/rvm/src/vm/libSupport/java/lang/reflect In directory sc8-pr-cvs6.sourceforge.net:/tmp/cvs-serv20207
Modified Files:
	VMArray.java
Log Message:
Fix to Ian's recent commit.  This re-wored fix was done by Ian.  I'm
committing it because I want it in before tonight's nightly regression.
From Ian's email:

"I believe this is both more efficient and equivalent to what the JNI
version of the code was doing."