MMTk/ext/vm/JikesRVM/org/mmtk/vm Collection.java,1.8,1.9 Update of /cvsroot/jikesrvm/MMTk/ext/vm/JikesRVM/org/mmtk/vm In directory sc8-pr-cvs6.sourceforge.net:/tmp/cvs-serv13453/MMTk/ext/vm/JikesRVM/org/mmtk/vm
Modified Files:
	Collection.java
Log Message:
Fixed bug in out-of-memory throwing code.

The code did not allow for the fact that the "reserved" pages may
(significantly) exceed the total available.  So the amount we grow the heap
needs to be the sum of our current availability and the headroom for the OOE.

Since this lead to infinite regress, this may be the cause of some of the
infamous GC hangs...  Let's see...