Initial check in of an Ant based build system. Initially the ant based build system will sit beside the existing build infrastructure until such a time where it is validated and known to work on all platforms that the previous build system worked on.
To use ant to build the system run the ant command in the base directory. The build should give directions on how to progress to next stage of build process at each step along the way. 

Known to work on ia32 and ppc32 linux and compile on ppc64 linux. Other platforms will be supported in the near future.

NOTE: The created classpath directory is structured differently from the way we currently do it so you should just download and rebuild accordingto the instructions the build file gives.