Need OBJECT_MODE=32_64 hack to get GNU classpath to build on 64bit AIX.
The problem is that ar and similar tools default to 32bit only mode and
need to be told to process 64 bit object files.  This can be done either
by setting the OBJECT_MODE environment variable or by trying to pass
through the appropriate command line arguments through the classpath 
configure step down to libtool, etc.  The environment variable is much simpler..