Probable fix for 1725306 (header corruption on Eclipse).
As reported by Daniel & Steve, the problem is that the
code in OPT_ExpandRuntimeServices in the opt compiler was
improperly getting a default value (null === 0) for the offset
of the thin lock word when the compile time type of the object
being synchronized on was an interface.

In the default object model, all objects have thin locks
and they are always at the same fixed offset.  However, the
code still has some support for other object models (in particular 
the ECOOP'02 paper by Bacon, Fink, and Grove) where the thin lock
may be at different offsets for different types.  

The real fix is in the VM_Class changes:
  (1) in resolve we set thinLockOffset to Offset.max() for interfaces
  (2) assert(isResolved()) in getThinLockOffset to catch us trying
      to look at thinLockOffset before it has a valid value.

I also did a minor cleanup by removing setThinLockOffset as an
abstract method of VM_Type and left it only as a method on VM_Class and
made it more explicit that arrays always have a  thinlock word, 
and it is always at the default offset.