Implementation of SSE2 for the opt compiler. The target.arch.sse2 option now has the value of none/base/full. When set to full, the opt compiler uses SSE2 instead of x87 for the majority of operations. This full mode also changes our internal calling convention to use the SSE2 registers for parameter passing and return values.
This is mostly a naive but (hopefully) correct implementation from which 
we can move forward.

Obviously this is a large change. I have tested on x86 and PPC and with 
the option on all three settings with commit tests and saw no 
regressions. I am kicking off another set now, and establishing an SSE2 
commit test to run daily to test this feature.