RVM-82: Exit VM immediately if hardware trap occurs in Uninterruptible code.
Changed deliverHardwareException to check if the IP of the faulting/trapping instruction is
in an uninterruptible method.  If it is, then we print a diagnostic and sysFail immediately. The hope of that this will result in a cleaner error then if we allow the system to stumble along for a little while longer before it dies.