Fix problems with 14133.
The extant bit on RegisterOperand already had a fairly precise meaning:  that the value is known to pre-exist (in the sense of Agesen and Detlefs '99 ECOOP paper) and that therefore we can do speculative class hierarchy based inlining without requiring a guard if the value was extant.  Extant does not imply non-null.  We could add a non-null bit, but it needs to be distinct from extant. 

Before we only set extant for the formal parameters of a method, but I believe the extension made in 14133 to new/newarray works as well as long as we restrict it to resolved news (if the new is resolved, then the class must already be loaded before the current invocation of the method).  I don't think this will actually help the inliner at all (since the preciseType bit is set on the results of all news anyways, and we should be devirtualizing based on the preciseType without guards), but it  is harmless to set the extant bit on the result of resolved news.