RVM-645 : Fix bug in LiveAnalysis.
There's a good chance this will fix a number of the GC crashes with the opt compiler.  Comment from JIRA appended.

The problem was that in LiveAnalysis.processBlock the code was calling getContainsPEIWithHandler() before it called computeBlockGenAndKill. Unfortunately, the value for getContainsPEIWithHandler is actually computed by computeBlockGenAndKill, so it was always returning false when it was called to soon. As a result, the code was treating all IN edges from exception edges as if they were IN edges from normal control flow. This resulted in a too aggressive killing of values that were live via exception edges.

The fix is simply to move the call to computeBlockGenAndKill to the top of processBlock, then all of the info about the block is computed in time and the code works as intended.